requirements.txt
# Core Python Version
python>=3.11,<3.12

# Web Framework & API
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-socketio>=5.9.0

# Data Processing & Analysis
numpy>=1.24.0
pandas>=2.1.0
matplotlib>=3.7.0
plotly>=5.17.0
scipy>=1.11.0

# System Monitoring & Hardware
psutil>=5.9.6
pynvml>=12.535.77
GPUtil>=1.4.0
nvidia-ml-py>=12.535.77

# HTTP & Networking
requests>=2.31.0
aiohttp>=3.9.0

# Configuration & CLI
typer>=0.9.0
pyyaml>=6.0.1
click>=8.1.0
python-dotenv>=1.0.0

# Logging & Output
rich>=13.6.0
loguru>=0.7.2
structlog>=23.2.0

# File System Monitoring
watchdog>=3.0.0

# Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.9.0
flake8>=6.1.0

# Windows-specific Dependencies
pywin32>=306; sys_platform == "win32"
wmi>=1.5.1; sys_platform == "win32"
